Fixed a bug that makes Grasshopper previews visible on Revit type preview dialog.
